feat: Introduce ManagedDataSource.example
fix: UI behavior in ExamView
